Remove Python 3.5, tests more on 3.6-3.7, update supported OSes
We have test issues on Python 3.5. Instead of fixing them,
as Python 3.5 is EOL and is being killed by pip by the end of
Jan 2021, we decided to remove support for it now.

Also start running integration tests on Python 3.6 and 3.7, along
with 3.8.

Explain 3.9 status in the README too.

Finally stop claiming OS independence as the app is tested only
on macOS and GNU/Linux.