Exit with 2 at the end if there were processing errors
(This is a breaking change because GitLabForm v1.x exited with 0
in such case!)

List the groups and projects that have failed, with their numbers.
Add --start-from-group to allow retries from a given group number.
(We used to have only --start-from for projects.)
Always print the stacktraces on processing errors.
Unify and document (in -h output) the exit codes used.

Fixes #153.