Add only users to project if they are not properly set
Fixes #101. Checks if the user is added with the correct access level and
expire date. If those values are not the ones defined by gitlabform
the user gets deleted if it's present and added again.

This commit also fixes #207, trying to use "expires_at" when setting
a user was making gitlabform to fail.