Better skip processing empty effective configs
for both groups and projects. Firstly we check if given entity
has a config that actually applies to it - f.e. if a group
has group-level configs that are processed by group processors.
Secondly we skip groups and projects with an empty effective
config at the beginning of processing and show that in the output.

Also document and refactor the code:
  * split ConfigurationProjectsAndGroups into
    ConfigurationProjectsAndGroups (with projects methods)
    and ConfigurationGroups (with groups methods),
  * move even more of the UI into ui.py,
  * simplify gitlabform.core thanks to the above
    and the fact that effective configs here are non-empty,
  * make debugging easier by adding main method to run.py,