Add validate break inheritance flag function
The merge_configs function was previously not checking if the break-inheritance flag is set at the proper level. The validation function validates if the break-inheritance flag is set at the top-level setting of the provided configuration. The validation function will quit the program if users set the break-inheritance flag improperly. To identify the setting levels more_general_config and more_specific_config refer to,  pass keyworded arguments to merge_configs.