Always expect pagination for GETs
as it happened a few times already that we didn't mark an API as using
pagination initially and it caused problems for some users. According
to GitLab's API docs (https://docs.gitlab.com/ee/api/#pagination)
the pagination "is available on all endpoints", so let's just pass
the for all GETs.

Fixes #354.