Fix bug with breaking config inheritance
not working for some configurations (#326).

Without this change, when I run the app on my config, using ALL_DEFINED,
at some point the self.configuration in the GitLabForm object gets changed:
"inherit: false" entries are gone and subsequent calls to the
ConfigurationProjectsAndGroups.get_effective_config_for_project()
return configs without breaking inheritance.

The bug has to be caused by removal if the "inherit: false" that we do
in replace_config_sections() because we have been manipulating *references*
to the single self.configuration here.

deepcopy() makes sure that we only change copies of it and fixes the bug
according to my test runs on my real config. Unfortunately I am not able
to reproduce the bug with a unit test.