Refactor: renames and moves for maintainability
Firstly, 'input' is now called 'lists' as input is too
unclear. Hopefully toghether with class names it's easier
to understand that this code operates on lists of groups
and/or projects to be configured.

However I can't find a better name for the file with
the side-effect, optional effective configuration file
creation code than 'output'...

Configuration transformation have joined the other
configuration-related code.

Most of the "main" code for running GitLabForm has been
moved directly into 'gitlabform' package. There is no
more 'ui.py' because that code was not only printing
out UI output, but also contains the logic.