feat: add json schema for top level gitlabform configuration syntax
The schemas can be submitted to schemastore, which is supported by
various editors. The can help guide users to write gitlabform
configuration by showing auto completion, hints, examples, or
description of a configuration.

Added necessary updates to docs as well as GitHub Actions CI, which
executes the unit tests for the schemas.