chore(dev/ci): set expiry date for access token
As of GitLab v16.0, all access tokens must have an expiry date. Otherwise the access token will not be created. This wasn't being done in the script that sets up a GitLab instances with an access token so that the token can be used for all acceptance tests via API calls. This worked fine before but started failing recently when latest version of GitLab was being used in the CI because the latest version is 16 and up. As a result, all the PR validations are failing because the acceptance tests are failing with '401 Unauthorized' error.