refactor: process project archive/unarchive using python-gitlab library
Moved the Gitlab object created from GitlabWrapper to the AbstractProcessor
class so that the object is available to all other processors. Then updated
project processor code to use the Gitlab object, which comes from the
python-gitlab library, for performing project archive and unarchive.

Removed the internal API/method that was previously used for calling the
GitLab's API for project archive and unarchive purpose.