fix: dry run failure when project transfer is configured along with other additional configs (#726)
The `project: transfer_from` config fails when run with `--noop`, if there are other configurations for the project (i.e. `project_settings`). This is due to the `project_settings` processor getting a 404 when trying to fetch the settings from GitLab API(since the project doesn't get moved in dry-run).

The dry-run/noop logic is updated to account for situation when project transfer configuration is used. Additional relevant test cases have been included as well.

Fixes #725