Release only when tests pass
and revampt the way we organize GitHub Actions
code.

The "_" YAMLs are the workflows we run for main branch,
in PRs and on release (tags beginning with "v").

The other YAMLs are being used in the above ones according
to "reusable workflows" pattern
(https://docs.github.com/en/actions/using-workflows/reusing-workflows).

It's not perfect, but I think it's easier to comprehend than
what we had before.