perf: reduce execution time of branch processor
Previously a fix was added so that members configuration is processed
before processing other configurations that are dependent on it. See
PR #857 for details.

This revealed an issue in GitLab. After a user is added to a project, if
that user is referenced in other processor like branch protection,
the processing doesn't take place successfully. In case of branch
protection an error is received saying the user is not a member of
the project. In other processors, there are no errors but the config
is not applied. The work-around for this GitLab issue was to wait a little.

The work-around mentioned above for waiting was added to branch
protection inside a loop so that it will wait for every branch config.
But that is not necessary. Waiting once at the beginning of branch
config processing should be sufficient. This will save a lot of time
when processing many projects with multiple branches for those projects.

related to #877