perf: reduce execution time due to project members config
Previously a change was relased in v4.0.1 that fixed the
ordering of config processor. One change included in it
increased the overall execution time of gitlabform if
the config contained members for a project. This was due
to acceptance tests failing when those users were being
referenced in other config sections. The change added
delays causing the execution time increase.

After further investigation, it seems that delay is only
needed in the acceptance tests because gitlabform does
not actually create users. The tests setup creates new
users and in that case a small delay is needed.

This change removes the unnecessary delay that was
included in the runtime of gitlabform. The delay has
been added to necessary tests setup.