docs: update ci_cd_variables.md to include raw variable example (#1091)
Gitlab's UI uses the term "expanded" when indicating whether "$ will be treated as the start of a reference to another variable" or not.  The API uses the term "raw" instead, which is confusing if you're not familiar with the API.  This adds an example, since the docs weren't clear they definitely supported this option.