test: fix group push rule test assertion
One of the test for group push rule was set to validate the
'commit_committer_check' attribute. Currently it is set to 'None'
if the attribute is not set. But in recent GitLab release, the
API is returning boolean value. So, the test is failing causing
the CI to be broken

This change updates test assertion to match with what GitLab's
API returns. This will restore the CI pipeline to working state.